Cloze button get disabled outside of cloze field
If the user tries to add a cloze in a field that is not a cloze type,
the user gets an alert message.

I could have removed the buttons or disabled them, as it's done when
the note type is not a cloze deletion, but I think it's better to
avoid changes in the buttons. And anyway, the alert would still be
useful if the user tried to use the shortcut to add the cloze.

In order to do this, I added a back-end method (that I expect we may
reuse in ankidroid) to get the index of the fields used in cloze. This
set is sent to the note editor, which propagates it where needed.

In mathjax, the cloze symbol is removed when the selected field is not
a cloze field

I would appreciate help, as the alert message is not properly
displayed, instead it shows

> missing key: adding-cloze-outside-cloze-field

+++ rslib/src/findreplace.rs
@@ -74,7 +74,7 @@ impl Collection {
                     None => FieldForNotetype::Any,
                     Some(name) => match nt.get_field_ord(name) {
                         None => FieldForNotetype::None,
-                        Some(ord) => FieldForNotetype::Index(ord),
+                        Some(ord) => FieldForNotetype::Index(ord as usize),
                     },
                 };
                 last_ntid = Some(nt.id);
